+++ packages/lesswrong/lib/abTestImpl.ts
@@ -1,4 +1,6 @@
-import React from 'react';
+import React, { useContext } from 'react';
+import { useCookies } from 'react-cookie'
+import { useCurrentUser } from '../components/common/withUser';
 import * as _ from 'underscore';
 import rng from './seedrandom';
 
@@ -126,4 +128,50 @@ function weightedRandomPick(options: Record<string,number>, seed: string): strin
       return key;
   }
   throw new Error("Out of range value in weightedRandomPick");
+}
+
+
+// Returns the name of the A/B test group that the current user/client is in.
+export function useABTest(abtest: ABTest): string {
+  const currentUser = useCurrentUser();
+  const clientId = useClientId();
+  const abTestGroups: RelevantTestGroupAllocation = useContext(ABTestGroupsContext);
+  const group = getUserABTestGroup(currentUser, clientId, abtest);
+  
+  abTestGroups[abtest.name] = group;
+  return group;
+}
+
+export function useABTestProperties(abtest: ABTest): ABTestGroup {
+  const groupName = useABTest(abtest);
+  return abtest.groups[groupName];
+}
+
+// Returns the user's clientID. This is stored in a cookie separately from
+// accounts; a user may have multiple clientIDs (eg if they have multiple
+// devices) and a clientID may correspond to multiple users (if they log out and
+// log in with a different account).
+//
+// A logged-out user's client ID determines which A/B test groups they are in.
+// A logged-in user has their A/B test groups determined by the client ID they
+// had when they created their account.
+export function useClientId(): string {
+  const [cookies] = useCookies(['clientId']);
+  return cookies.clientId;
+}
+
+// Return a complete mapping of A/B test names to A/B test groups. This is used
+// on the page that shows you what A/B tests you're in; it should otherwise be
+// avoided, since it interferes with caching.
+export function useAllABTests(): CompleteTestGroupAllocation {
+  const currentUser = useCurrentUser();
+  const clientId = useClientId();
+  
+  const abTestGroups: CompleteTestGroupAllocation = useContext(ABTestGroupsContext);
+  
+  const testGroups = getAllUserABTestGroups(currentUser, clientId);
+  for (let abTestKey in testGroups)
+    abTestGroups[abTestKey] = testGroups[abTestKey];
+  
+  return testGroups;
 }
