+++ packages/lesswrong/lib/index.ts
@@ -182,7 +182,6 @@ import './alignment-forum/users/views';
 
 import './analyticsEvents';
 import './abTests';
-import './abTestUtil';
 import './vulcan-i18n-en-us';
 
 import '../components/spuriousChange';
